+++ include/can/firmware/hal_can_bus.hpp
@@ -6,7 +6,11 @@
 #include "can/core/ids.hpp"
 #include "can/core/parse.hpp"
 #include "common/core/freertos_synchronization.hpp"
+#pragma GCC diagnostic push
+// NOLINTNEXTLINE(clang-diagnostic-unknown-warning-option)
+#pragma GCC diagnostic ignored "-Wvolatile"
 #include "platform_specific_hal_fdcan.h"
+#pragma GCC diagnostic pop
 
 namespace hal_can_bus {
