+++ include/common/firmware/spi_comms.hpp
@@ -2,7 +2,11 @@
 #include <cstdint>
 #include <span>
 
+#pragma GCC diagnostic push
+// NOLINTNEXTLINE(clang-diagnostic-unknown-warning-option)
+#pragma GCC diagnostic ignored "-Wvolatile"
 #include "platform_specific_hal_conf.h"
+#pragma GCC diagnostic pop
 
 namespace spi {
 struct SPI_interface {
