+++ pipettes/firmware/can_task.cpp
@@ -10,10 +10,8 @@
 #include "can/core/message_handlers/move_group_executor.hpp"
 #include "can/core/message_writer.hpp"
 #include "can/core/messages.hpp"
-#include "can/firmware/hal_can_message_buffer.hpp"
 #include "common/core/freertos_message_queue.hpp"
 #include "common/core/freertos_task.hpp"
-#include "common/firmware/can.h"
 #include "common/firmware/errors.h"
 #include "common/firmware/i2c_comms.hpp"
 #include "common/firmware/spi_comms.hpp"
@@ -22,7 +20,14 @@
 #include "motor-control/core/motor_messages.hpp"
 #include "pipettes/core/eeprom.hpp"
 #include "pipettes/core/message_handlers/eeprom.hpp"
+#pragma GCC diagnostic push
+// NOLINTNEXTLINE(clang-diagnostic-unknown-warning-option)
+#pragma GCC diagnostic ignored "-Wvolatile"
+#include "can/firmware/hal_can_message_buffer.hpp"
+#include "common/firmware/can.h"
+#include "motor_hardware.h"
 #include "platform_specific_hal_conf.h"
+#pragma GCC diagnostic pop
 
 using namespace hal_can_bus;
 using namespace can_messages;
@@ -47,85 +52,6 @@ static freertos_message_queue::FreeRTOSMessageQueue<Move> motor_queue(
     "Motor Queue");
 static freertos_message_queue::FreeRTOSMessageQueue<Ack> complete_queue(
     "Complete Queue");
-
-void HAL_SPI_MspInit(SPI_HandleTypeDef* hspi) {
-    GPIO_InitTypeDef GPIO_InitStruct = {0};
-    if (hspi->Instance == SPI2) {
-        /* Peripheral clock enable */
-        __HAL_RCC_SPI2_CLK_ENABLE();
-        __HAL_RCC_GPIOA_CLK_ENABLE();
-        __HAL_RCC_GPIOB_CLK_ENABLE();
-        /**SPI2 GPIO Configuration
-        PC6     ------> SPI2_CS
-        PB13     ------> SPI2_SCK
-        PB14     ------> SPI2_CIPO
-        PB15     ------> SPI2_COPI
-
-         Step/Dir
-         PC3  ---> Dir Pin
-         PC7  ---> Step Pin
-         Enable
-         PC8  ---> Enable Pin
-        */
-        GPIO_InitStruct.Pin = GPIO_PIN_13 | GPIO_PIN_14 | GPIO_PIN_15;
-        GPIO_InitStruct.Mode = GPIO_MODE_AF_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-        GPIO_InitStruct.Alternate = GPIO_AF5_SPI2;
-        HAL_GPIO_Init(GPIOB, &GPIO_InitStruct);
-
-        // Chip select
-        GPIO_InitStruct.Pin = GPIO_PIN_6;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-        // Enable/Dir/Step pin
-        GPIO_InitStruct.Pin = GPIO_PIN_3 | GPIO_PIN_7 | GPIO_PIN_8;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        GPIO_InitStruct.Pull = GPIO_NOPULL;
-        GPIO_InitStruct.Speed = GPIO_SPEED_FREQ_LOW;
-        HAL_GPIO_Init(GPIOC, &GPIO_InitStruct);
-
-        GPIO_InitStruct.Pin = GPIO_PIN_7 | GPIO_PIN_8;
-        GPIO_InitStruct.Mode = GPIO_MODE_OUTPUT_PP;
-        HAL_GPIO_Init(GPIOC,  // NOLINT(cppcoreguidelines-pro-type-cstyle-cast)
-                      &GPIO_InitStruct);
-    }
-}
-SPI_HandleTypeDef hspi2 = {
-    .Instance = SPI2,
-    .Init = {.Mode = SPI_MODE_MASTER,
-             .Direction = SPI_DIRECTION_2LINES,
-             .DataSize = SPI_DATASIZE_8BIT,
-             .CLKPolarity = SPI_POLARITY_HIGH,
-             .CLKPhase = SPI_PHASE_2EDGE,
-             .NSS = SPI_NSS_SOFT,
-             .BaudRatePrescaler = SPI_BAUDRATEPRESCALER_32,
-             .FirstBit = SPI_FIRSTBIT_MSB,
-             .TIMode = SPI_TIMODE_DISABLE,
-             .CRCCalculation = SPI_CRCCALCULATION_DISABLE,
-             .CRCPolynomial = 7,
-             .CRCLength = SPI_CRC_LENGTH_DATASIZE,
-             .NSSPMode = SPI_NSS_PULSE_DISABLE}
-
-};
-
-void HAL_SPI_MspDeInit(SPI_HandleTypeDef* hspi) {
-    if (hspi->Instance == SPI2) {
-        /* Peripheral clock disable */
-        __HAL_RCC_SPI2_CLK_DISABLE();
-
-        /**SPI2 GPIO Configuration
-        PB12     ------> SPI2_NSS
-        PB13     ------> SPI2_SCK
-        PB14     ------> SPI2_MISO
-        PB15     ------> SPI2_MOSI
-        */
-        HAL_GPIO_DeInit(GPIOB,
-                        GPIO_PIN_12 | GPIO_PIN_13 | GPIO_PIN_14 | GPIO_PIN_15);
-        HAL_GPIO_DeInit(GPIOA, GPIO_PIN_8 | GPIO_PIN_9);
-    }
-}
 
 spi::SPI_interface SPI_intf = {
 
@@ -213,8 +139,8 @@ static auto dispatcher = Dispatcher(
     if (MX_FDCAN1_Init(&fdcan1) != HAL_OK) {
         Error_Handler();
     }
-    __HAL_RCC_SPI2_CLK_ENABLE();
-    if (HAL_SPI_Init(&hspi2) != HAL_OK) {
+
+    if (initialize_spi() != HAL_OK) {
         Error_Handler();
     }
     can_bus::setup_node_id_filter(can_bus_1, NodeId::pipette);
